moving the zod schema to utils file
