+++ src/tools/search.ts
@@ -5,11 +5,10 @@ import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { WebApi } from "azure-devops-node-api";
 import { IGitApi } from "azure-devops-node-api/GitApi.js";
 import { z } from "zod";
-import { apiVersion } from "../utils.js";
+import { apiVersion, stringArrayParam } from "../utils.js";
 import { orgName } from "../index.js";
 import { VersionControlRecursionType } from "azure-devops-node-api/interfaces/GitInterfaces.js";
 import { GitItem } from "azure-devops-node-api/interfaces/GitInterfaces.js";
-import { stringArrayParam } from "../shared/schemas.js";
 
 const SEARCH_TOOLS = {
   search_code: "search_code",
