+++ src/utils.ts
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { z } from "zod";
+
 export const apiVersion = "7.2-preview.1";
 export const batchApiVersion = "5.0";
 export const markdownCommentsApiVersion = "7.2-preview.4";
@@ -88,3 +90,13 @@ export function streamToString(stream: NodeJS.ReadableStream): Promise<string> {
     stream.on("end", () => resolve(data));
   });
 }
+
+/**
+ * Creates an optional Zod schema that accepts either a single string or an array of strings,
+ * normalizing the input to always produce a string array.
+ */
+export const stringArrayParam = (description: string) =>
+  z
+    .union([z.string().transform((value) => [value]), z.array(z.string())])
+    .optional()
+    .describe(description);
